Use a AbstractTreeGrower
